create private_functions file, exporting constants, format tests, .gitignore
