+++ .gitignore
@@ -7,8 +7,6 @@ javascript/node_modules
 ruby/*.gem
 ruby/.rspec_status
 
-javascript/node_modules
-
 ruby/coverage
 
 # Rust crate
