+++ javascript/lib/do_username.js
@@ -1,37 +1,8 @@
-const {
-    SEA_LIST,
-    SEA_CREATURES,
-    DESCRIPTORS,
-    COLORS,
-    CREATURE_DESCRIPTORS,
-} = require('./constants');
+const { SEA_LIST, SEA_CREATURES, DESCRIPTORS, COLORS, CREATURE_DESCRIPTORS } = require('./constants');
 
-const random = (items) => items[Math.random() * items.length | 0];
+const { randomNoun, randomDescriptor, randomColor, format, combineUsername } = require('./private_functions');
 
-const randomNoun = module.exports.randomNoun = () => random(SEA_LIST);
-
-const randomDescriptor = module.exports.randomDescriptor = (noun) => {
-    if (!SEA_CREATURES.includes(noun)) return random(DESCRIPTORS);
-
-    return random(DESCRIPTORS.concat(CREATURE_DESCRIPTORS));
-};
-
-const format = module.exports.format = (string) => string[0].toUpperCase() + string.slice(1);
-
-const randomColor = module.exports.randomColor = () => random(COLORS);
-
-const combineUsername = module.exports.combineUsername = (maxSize, descriptor, color, noun) => {
-    if ((descriptor + color + noun).length <= maxSize)
-        return descriptor + color + noun;
-    else if ((descriptor + noun).length <= maxSize)
-        return descriptor + noun;
-    else if ((color + noun).length <= maxSize)
-        return color + noun;
-    else
-        return noun.slice(0, maxSize);
-};
-
-module.exports.generate = (maxSize = 30) => {
+const generate = (maxSize = 30) => {
     if (maxSize <= 0 || typeof maxSize !== 'number') {
         throw new Error('The maxSize argument must be an integer number greater than zero.');
     }
@@ -45,4 +16,13 @@ module.exports.generate = (maxSize = 30) => {
     color = format(color);
 
     return combineUsername(maxSize, descriptor, color, noun);
+};
+
+module.exports = {
+    SEA_LIST,
+    SEA_CREATURES,
+    DESCRIPTORS,
+    COLORS,
+    CREATURE_DESCRIPTORS,
+    generate,
 };
